+++ src/import.ts
@@ -3,6 +3,7 @@ import "./border.css";
 import "./breakpoint.css";
 import "./breakpoint.js";
 import "./color.css";
+import "./color.js";
 import "./filter.css";
 import "./font.css";
 import "./opacity.css";